+++ sources/quick_maths/vector_x_scalar.c
@@ -12,7 +12,7 @@
 
 #include "../../includes/minirt.h"
 
-t_vector vector_x_scalar(t_vector vector, float scalar)
+t_vector vector_x_scalar(t_vector vector, double scalar)
 {
 	t_vector result;
 
@@ -22,9 +22,9 @@ t_vector vector_x_scalar(t_vector vector, float scalar)
 	return (result);
 }
 
-float vector_scalar(t_vector vector1, t_vector vector2)
+double vector_scalar(t_vector vector1, t_vector vector2)
 {
-	float result;
+	double result;
 
 	result = (vector1.x * vector2.x) + (vector1.y * vector2.y) + (vector1.z * vector2.z);
 	return (result);
